refactor: separate each nav into its comp
